separate the sideitem code to a file and add the icons to the project
